Combined cpp files
Combined the code in Quarterback.cpp, Center.cpp and Kicker.cpp into QB-Cent-Kick.cpp. This file will use the EEPROM check similar to the way it used to be done to check the bot type, this revision is extremely rough, just trying to convey the idea
