+++ platformio.ini
@@ -22,7 +22,6 @@ lib_deps =
 board = uno
 build_src_filter = 
 	+<main.cpp>
-	+<ramp.cpp>
 	+<PolarRobotics.cpp>
 	+<PolarRobotics.h>
 
@@ -31,8 +30,10 @@ board = megaatmega2560
 build_src_filter = 
 	-<main.cpp>
 	-<test.cpp>
+	+<QB-Cent-Kick.cpp>
+	+<PolarRobotics.cpp>
+	+<PolarRobotics.h>
 
 [env:old_linemen_working]
 board = uno
-build_src_filter = 
-	+<test.cpp>
+build_src_filter = +<test.cpp>
